Реалізовано базові таблиці та ендпоінти microsoft asp identity. Видалено userId із заголовку task endpoints.

+++ TodoApp.WebApi/Endpoints/TaskEndpoints.cs
@@ -1,3 +1,4 @@
+using System.Security.Claims;
 using Microsoft.AspNetCore.Mvc;
 using TodoApp.Services.DTOs;
 using TodoApp.Services.Interfaces;
@@ -8,22 +9,38 @@ public static class TaskEndpoints
 {
     public static void MapTaskEndpoints(this IEndpointRouteBuilder app)
     {
-        var group = app.MapGroup("tasks").WithTags("Tasks");
+        var group = app.MapGroup("tasks").WithTags("Tasks")
+        .RequireAuthorization();;
 
-        group.MapGet("/", async (Guid userId, ITaskService taskService) =>
+        group.MapGet("/", async (ClaimsPrincipal user, ITaskService taskService) =>
         {
+            var userIdClaim = user.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if(!Guid.TryParse(userIdClaim, out var userId))
+                return Results.Unauthorized();
+
             var tasks = await taskService.GetAllAsync(userId);
             return Results.Ok(tasks);
         });
 
-        group.MapGet("/{id:guid}", async (Guid id, Guid userId, ITaskService taskService) =>
+        group.MapGet("/{id:guid}", async (Guid id, ClaimsPrincipal user, ITaskService taskService) =>
         {
+            var userIdClaim = user.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if(!Guid.TryParse(userIdClaim, out var userId))
+                return Results.Unauthorized();
+
             var task = await taskService.GetByIdAsync(id, userId);
             return Results.Ok(task);
         }).WithName("GetTaskById");
 
-        group.MapPost("/", async ([FromBody] CreateTodoTaskDto dto, Guid userId, ITaskService taskService) =>
+        group.MapPost("/", async ([FromBody] CreateTodoTaskDto dto, ClaimsPrincipal user, ITaskService taskService) =>
         {
+            var userIdClaim = user.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if(!Guid.TryParse(userIdClaim, out var userId))
+                return Results.Unauthorized();
+
             if (string.IsNullOrEmpty(dto.Title))
                 return Results.BadRequest(new {message = "Назва завдання не може бути попрожнім"});
 
@@ -33,20 +50,35 @@ public static class TaskEndpoints
 
         });
         
-        group.MapPut("/{id:guid}", async (Guid id, [FromBody] CreateTodoTaskDto dto, Guid userId, ITaskService taskService) =>
+        group.MapPut("/{id:guid}", async (Guid id, [FromBody] CreateTodoTaskDto dto, ClaimsPrincipal user, ITaskService taskService) =>
         {
+            var userIdClaim = user.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if(!Guid.TryParse(userIdClaim, out var userId))
+                return Results.Unauthorized();
+
             var updated = await taskService.UpdateAsync(id, dto, userId);
             return updated ? Results.NoContent() : Results.NotFound(new {message = "Завдання не знайдено або доступ заборонено"});
         });
 
-        group.MapDelete("/{id:guid}", async (Guid id, Guid userId, ITaskService taskService) =>
+        group.MapDelete("/{id:guid}", async (Guid id, ClaimsPrincipal user, ITaskService taskService) =>
         {
+            var userIdClaim = user.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if(!Guid.TryParse(userIdClaim, out var userId))
+                return Results.Unauthorized();
+
            var deleted =  await taskService.DeleteAsync(id, userId);
            return deleted ? Results.NoContent() : Results.NotFound(new { message = "Завдання не знайдено або доступ заборонено"}); 
         });
 
-        group.MapPatch("/{id:guid}/toggle", async (Guid id, Guid userId, ITaskService taskService) =>
+        group.MapPatch("/{id:guid}/toggle", async (Guid id, ClaimsPrincipal user, ITaskService taskService) =>
         {
+            var userIdClaim = user.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if(!Guid.TryParse(userIdClaim, out var userId))
+                return Results.Unauthorized();
+
             var updated = await taskService.ToggleCompleteAsync(id, userId);
             return updated ? Results.NoContent() : Results.NotFound(new {message = "Завдання не знайдено або доступ заборонено"});
         });
